moved functions creating backend test files/folders/package.json jest config into seperate file
